extract the logic that isn't shared and move it to their respective owners
