+++ common/src/main/scala/providers/Error.scala
@@ -1,4 +1,4 @@
-package notifications.providers
+package providers
 
 trait Error {
   def providerName: String
