+++ notification/app/notification/services/NotificationSenderSupport.scala
@@ -3,7 +3,6 @@ package notification.services
 import javax.inject.Inject
 
 import azure.{NotificationHubClient, NotificationHubConnection}
-import notifications.providers.{NotificationSender, WindowsNotificationProvider}
 import play.api.libs.ws.WSClient
 
 import scala.concurrent.ExecutionContext
@@ -17,5 +16,5 @@ final class NotificationSenderSupport @Inject()(wsClient: WSClient, configuratio
 
   private val hubClient = new NotificationHubClient(hubConnection, wsClient)
 
-  val notificationSender: NotificationSender = new WindowsNotificationProvider(hubClient)
+  val notificationSender: NotificationSender = new WindowsNotificationSender(hubClient)
 }
