+++ registration/app/registration/services/NotificationRegistrarSupport.scala
@@ -5,7 +5,6 @@ import javax.inject.Inject
 import azure.{NotificationHubClient, NotificationHubConnection}
 import com.google.inject.ImplementedBy
 import models.{Registration, WindowsMobile}
-import notifications.providers.{NotificationRegistrar, WindowsNotificationProvider}
 import play.api.libs.ws.WSClient
 
 import scala.concurrent.ExecutionContext
@@ -27,7 +26,7 @@ final class NotificationRegistrarSupport @Inject()(wsClient: WSClient, configura
 
   private val hubClient = new NotificationHubClient(hubConnection, wsClient)
 
-  private lazy val notificationRegistrar: NotificationRegistrar = new WindowsNotificationProvider(hubClient)
+  private lazy val notificationRegistrar: NotificationRegistrar = new WindowsNotificationRegistrar(hubClient)
 
   override def registrarFor(registration: Registration): \/[String, NotificationRegistrar] = registration match {
     case Registration(_, WindowsMobile, _, _) => notificationRegistrar.right
